requirements.txt
amqp==2.5.2
attrs==19.3.0
bcrypt==3.1.7
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.0.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
decorator==4.4.2
docutils==0.16
firebase-admin==4.0.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Testing==0.8.0
future==0.18.2
geocoder==1.38.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
more-itertools==8.2.0
msgpack==1.0.0
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.3
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.6
pytest==5.4.1
python-geohash==0.8.5
pytz==2019.3
ratelim==0.1.6
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.9
Werkzeug==1.0.0
zipp==3.1.0
